arreglos y nuevas funcionalidades

+++ backend/src/modules/products/products.service.ts
@@ -4,7 +4,8 @@ import { UpdateProductDto } from './dto/update-product.dto'
 import { ProductResponseDto } from './dto/product-response.dto'
 import { InjectRepository } from '@nestjs/typeorm'
 import { Product } from './entities/product.entity'
-import { Repository } from 'typeorm'
+import { Repository, ILike } from 'typeorm'
+import { CATEGORIES } from 'src/constants/categories.enum'
 
 @Injectable()
 export class ProductsService {
@@ -34,7 +35,10 @@ export class ProductsService {
 
       return {
         status: HttpStatus.OK,
-        message: 'ok',
+        message:
+          products.length < 1
+            ? 'No hay productos disponibles en este momento'
+            : 'ok',
         data: products,
       }
     } catch (error) {
@@ -42,13 +46,63 @@ export class ProductsService {
     }
   }
 
+  async findByMatch(filter: string) {
+    const products = await this.productRepository.find({
+      where: [
+        { name: ILike(`%${filter}%`) },
+        { description: ILike(`%${filter}%`) },
+      ],
+    })
+
+    return {
+      status: HttpStatus.OK,
+      message:
+        products.length < 1
+          ? `No se encontraron productos disponibles que incluyan la palabra ${filter}`
+          : 'ok',
+      data: products,
+    }
+  }
+
   async findOne(id: number): Promise<ProductResponseDto> {
     try {
       const product = await this.productRepository.findOne({ where: { id } })
+
+      if (product === null)
+        throw new HttpException(
+          `El producto con id ${id} ha sido eliminado`,
+          HttpStatus.BAD_REQUEST,
+        )
+
       return {
         status: HttpStatus.OK,
         message: 'ok',
         data: product,
+      }
+    } catch (error) {
+      throw new HttpException(error, HttpStatus.BAD_REQUEST)
+    }
+  }
+
+  async getCategories() {
+    return {
+      status: HttpStatus.OK,
+      message: 'ok',
+      data: CATEGORIES,
+    }
+  }
+
+  async findByCategory(category: CATEGORIES): Promise<ProductResponseDto> {
+    try {
+      const products = await this.productRepository.findBy({ category })
+
+      return {
+        status: HttpStatus.OK,
+        message:
+          products.length < 1
+            ? `no se encontraron productos de categoria ${category}`
+            : 'ok',
+        data: products,
       }
     } catch (error) {
       throw new HttpException(error, HttpStatus.BAD_REQUEST)
@@ -60,8 +114,15 @@ export class ProductsService {
     updateProductDto: UpdateProductDto,
   ): Promise<ProductResponseDto> {
     try {
-      await this.productRepository.update(id, updateProductDto)
       const product = await this.productRepository.findOne({ where: { id } })
+
+      if (product === null || !product)
+        throw new HttpException(
+          `El producto con id ${id} no existe o ha sido eliminado`,
+          HttpStatus.BAD_REQUEST,
+        )
+
+      await this.productRepository.update(id, updateProductDto)
       return {
         status: HttpStatus.OK,
         message: 'update',
@@ -75,6 +136,13 @@ export class ProductsService {
   async remove(id: number): Promise<ProductResponseDto> {
     try {
       const product = await this.productRepository.findOne({ where: { id } })
+
+      if (product === null || !product)
+        throw new HttpException(
+          `El producto con id ${id} no existe o ya ha sido eliminado`,
+          HttpStatus.BAD_REQUEST,
+        )
+
       await this.productRepository.remove(product)
       return {
         status: HttpStatus.OK,
